avoid unused parameter warning
git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@3265 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/sdpa/trunk/tests/sdpa/test_LoadBalancer.cpp
@@ -76,7 +76,7 @@ void LoadBalancerTest::testLoadBalancer()
 		try {
 			const Worker::ptr_t& ptrCurrWorker = wm.getNextWorker();
 
-			for( int l=0; l<arrLoad[k]; l++)
+			for( size_t l=0; l<arrLoad[k]; l++)
 			{
 				sdpa::job_id_t jobId;
 				ptrCurrWorker->dispatch(jobId);
